feat: redo hour format and smaller changes
